requirements.txt
reflex
supabase
python-dotenv
pydantic
